refactor(app): isolate media commands
